# BackendSpring/src/main/java/heg/backendspring/api/ProcessingApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.9.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package heg.backendspring.api;

import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-23T19:52:08.287321300+01:00[Europe/Zurich]", comments = "Generator version: 7.9.0")
@Validated
@Tag(name = "processing", description = "Traitement des rendus et gestion avancée des TPs")
public interface ProcessingApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /course/{courseId}/manageTP/{tpNo} : Gestion avancée d’un TP.
     * Effectue des opérations de gestion avancée sur un TP.
     *
     * @param courseId Identifiant du cours. (required)
     * @param tpNo Numéro de TP. (required)
     * @return Résultat de la gestion du TP. (status code 200)
     *         or Cours ou TP introuvable. (status code 404)
     *         or Erreur interne. (status code 500)
     */
    @Operation(
        operationId = "manageTP",
        summary = "Gestion avancée d’un TP.",
        description = "Effectue des opérations de gestion avancée sur un TP.",
        tags = { "processing" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Résultat de la gestion du TP.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = heg.backendspring.models.TPDto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Cours ou TP introuvable."),
            @ApiResponse(responseCode = "500", description = "Erreur interne.")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/course/{courseId}/manageTP/{tpNo}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<heg.backendspring.models.TPDto> manageTP(
        @Parameter(name = "courseId", description = "Identifiant du cours.", required = true, in = ParameterIn.PATH) @PathVariable("courseId") Long courseId,
        @Parameter(name = "tpNo", description = "Numéro de TP.", required = true, in = ParameterIn.PATH) @PathVariable("tpNo") Integer tpNo
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : 10, \"false\" : 1, \"courseId\" : 1, \"submission\" : { \"id\" : 100, \"fileName\" : \"TP1_Rendus.zip\", \"pathStorage\" : \"/data/courses/61-13/TP1/TP1_Rendus.zip\", \"pathFileStructured\" : \"/data/courses/61-13/TP1/TP1_RenduRestructuration.zip\" }, \"statusStudents\" : [ { \"id\" : 1000, \"studentId\" : 300, \"tpId\" : 10, \"studentSubmission\" : true } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /course/{courseId}/startProcessSubmission/{tpNo} : Lancement du traitement des rendus.
     * Lance le processus de restructuration et de traitement des rendus pour un TP.
     *
     * @param courseId Identifiant du cours. (required)
     * @param tpNo Numéro de TP. (required)
     * @return Traitement lancé. (status code 200)
     *         or Cours ou TP introuvable. (status code 404)
     *         or Erreur interne. (status code 500)
     */
    @Operation(
        operationId = "startProcessSubmission",
        summary = "Lancement du traitement des rendus.",
        description = "Lance le processus de restructuration et de traitement des rendus pour un TP.",
        tags = { "processing" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Traitement lancé.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = heg.backendspring.models.TPDto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Cours ou TP introuvable."),
            @ApiResponse(responseCode = "500", description = "Erreur interne.")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/course/{courseId}/startProcessSubmission/{tpNo}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<heg.backendspring.models.TPDto> startProcessSubmission(
        @Parameter(name = "courseId", description = "Identifiant du cours.", required = true, in = ParameterIn.PATH) @PathVariable("courseId") Long courseId,
        @Parameter(name = "tpNo", description = "Numéro de TP.", required = true, in = ParameterIn.PATH) @PathVariable("tpNo") Integer tpNo
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : 10, \"false\" : 1, \"courseId\" : 1, \"submission\" : { \"id\" : 100, \"fileName\" : \"TP1_Rendus.zip\", \"pathStorage\" : \"/data/courses/61-13/TP1/TP1_Rendus.zip\", \"pathFileStructured\" : \"/data/courses/61-13/TP1/TP1_RenduRestructuration.zip\" }, \"statusStudents\" : [ { \"id\" : 1000, \"studentId\" : 300, \"tpId\" : 10, \"studentSubmission\" : true } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}