# BackendSpring/src/main/resources/openapi/openapi.yaml
...
openapi: 3.0.4
info:
  title: AssistantIO API
  description: API de gestion des cours, TPs, étudiants et rendus.
  version: 1.0.0

# ==============================
#   SERVERS
# ==============================
servers:
  - url: http://localhost:8080
    description: Environnement local

# ==============================
#   TAGS
# ==============================
...
paths:
...
  /course/{courseId}/startProcessSubmission/{tpNo}:
    post:
      tags: [processing]
      summary: Lancement du traitement des rendus.
      description: Lance le processus de restructuration et de traitement des rendus pour un TP.
      operationId: startProcessSubmission
      parameters:
        - name: courseId
          in: path
          required: true
          description: Identifiant du cours.
          schema:
            type: integer
            format: int64
        - name: tpNo
          in: path
          required: true
          description: Numéro de TP.
          schema:
            type: integer
      responses:
        '200':
          description: Traitement lancé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TPDto'
        '404':
          description: Cours ou TP introuvable.
        '500':
          description: Erreur interne.
  /course/{courseId}/manageTP/{tpNo}:
    post:
      tags: [processing]
      summary: Gestion avancée d’un TP.
      description: Effectue des opérations de gestion avancée sur un TP.
      operationId: manageTP
      parameters:
        - name: courseId
          in: path
          required: true
          description: Identifiant du cours.
          schema:
            type: integer
            format: int64
        - name: tpNo
          in: path
          required: true
          description: Numéro de TP.
          schema:
            type: integer
      responses:
        '200':
          description: Résultat de la gestion du TP.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TPDto'
        '404':
          description: Cours ou TP introuvable.
        '500':
          description: Erreur interne.


# ==============================
#   SCHEMAS (DTO)
# ==============================
components:
  schemas:
...
    SubmissionDto:
      type: object
      description: Représentation d’un rendu (fichier ZIP) pour un TP.
      properties:
        id:
          type: integer
          format: int64
          example: 100
        fileName:
          type: string
          example: "TP1_Rendus.zip"
        pathStorage:
          type: string
          description: Chemin de stockage du ZIP original.
          example: "/data/courses/61-13/TP1/TP1_Rendus.zip"
        pathFileStructured:
          type: string
          description: Chemin du ZIP restructuré.
          example: "/data/courses/61-13/TP1/TP1_RenduRestructuration.zip"
      required: ["fileName", "pathStorage"]
      example:
        id: 100
        fileName: "TP1_Rendus.zip"
        pathStorage: "/data/courses/61-13/TP1/TP1_Rendus.zip"
        pathFileStructured: "/data/courses/61-13/TP1/TP1_RenduRestructuration.zip"
    TPDto:
      type: object
      description: Représentation d’un TP pour un cours.
      properties:
        id:
          type: integer
          format: int64
          example: 10
        "no":
          type: integer
          example: 1
        courseId:
          type: integer
          format: int64
          example: 1
        submission:
          $ref: '#/components/schemas/SubmissionDto'
        statusStudents:
          type: array
          items:
            $ref: '#/components/schemas/TPStatusDto'
      required: ["no", "courseId"]
      example:
        id: 10
        no: 1
        courseId: 1
        submission:
          id: 100
          fileName: "TP1_Rendus.zip"
          pathStorage: "/data/courses/61-13/TP1/TP1_Rendus.zip"
          pathFileStructured: "/data/courses/61-13/TP1/TP1_RenduRestructuration.zip"
        statusStudents:
          - id: 1000
            studentId: 300
            tpId: 10
            studentSubmission: true
    TPStatusDto:
      type: object
      description: Statut d’un étudiant pour un TP.
      properties:
        id:
          type: integer
          format: int64
          example: 1000
        studentId:
          type: integer
          format: int64
          example: 300
        tpId:
          type: integer
          format: int64
          example: 10
        studentSubmission:
          type: boolean
          example: true
      required: ["studentId", "tpId", "studentSubmission"]
      example:
        id: 1000
        studentId: 300
        tpId: 10
        studentSubmission: true